refactor: initiate splitting `StringUtil`
