+++ tools/src/main/java/org/aya/util/error/SourceFile.java
@@ -5,7 +5,6 @@ package org.aya.util.error;
 import com.intellij.openapi.util.text.Strings;
 import kala.collection.immutable.primitive.ImmutableIntArray;
 import kala.control.Option;
-import org.aya.util.StringUtil;
 import org.jetbrains.annotations.NotNull;
 
 import java.io.IOException;
@@ -23,7 +22,6 @@ public record SourceFile(
   @NotNull String sourceCode,
   @NotNull ImmutableIntArray lineOffsets
 ) {
-
   public static @NotNull SourceFile from(@NotNull SourceFileLocator locator, @NotNull Path path) throws IOException {
     return from(locator, path, Files.readString(path));
   }
@@ -33,15 +31,17 @@ public record SourceFile(
   }
 
   public SourceFile(@NotNull String display, @NotNull Option<Path> underlying, @NotNull String sourceCode) {
-    this(display, underlying, Strings.convertLineSeparators(sourceCode), StringUtil.indexedLines(sourceCode));
+    this(display, underlying, Strings.convertLineSeparators(sourceCode), ParsingUtil.indexedLines(sourceCode));
   }
 
   public SourceFile(@NotNull String display, @NotNull Path underlying, @NotNull String sourceCode) {
     this(display, Option.some(underlying), Strings.convertLineSeparators(sourceCode));
   }
 
-  public static final SourceFile NONE = new SourceFile("<unknown-file>", Option.none(), "");
-  public static final SourceFile SER = new SourceFile("<serialized-core>", Option.none(), "");
+  public static final SourceFile NONE =
+    new SourceFile("<unknown-file>", Option.none(), "", ImmutableIntArray.empty());
+  public static final SourceFile SER =
+    new SourceFile("<serialized-core>", Option.none(), "", ImmutableIntArray.empty());
 
   public boolean isSomeFile() {
     return underlying.isDefined();
